added EventStoreSession, renamed reader/writer
